requirements.txt
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
tabulate>=0.9.0
python-dotenv==1.1.1
requests==2.31.0
tqdm>=4.64.0
pyyaml>=6.0
openai>=1.0.0
httpx>=0.24.0
streamlit==1.28.2
matplotlib>=3.6.0
plotly==5.17.0
scipy>=1.10.0
jinja2>=3.1.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
torch==2.1.2
imbalanced-learn>=0.10.0
ta==0.10.2
kiteconnect>=4.0.0
schedule>=1.2.0
statsmodels>=0.14.0
psutil>=5.9.0
pydantic>=2.0.0
supabase>=2.0.0
asyncio-mqtt>=0.13.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
yfinance>=0.2.0
websocket-client>=1.6.0
redis>=5.0.0
celery>=5.3.0
schedule>=1.2.0
APScheduler>=3.10.0
python-dateutil>=2.8.0